+++ backend/src/main/java/entities/User.java
@@ -2,8 +2,6 @@ package entities;
 
 import com.googlecode.objectify.annotation.Entity;
 import com.googlecode.objectify.annotation.Id;
-
-import javax.jws.soap.SOAPBinding;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 
@@ -14,16 +12,16 @@ import javax.persistence.GenerationType;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String username;
     private String password;
     private String email;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
